requirements.txt
alembic==1.8.0
awscli==1.22.26
botocore==1.23.26
certifi==2022.5.18.1
click==8.1.3
colorama==0.4.3
dnspython==2.2.1
docutils==0.15.2
email-validator==1.2.1
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
is-disposable-email==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.38
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
wtforms-validators==1.0.0
zipp==3.8.0
